MKLP and others Counting script
This is version three of the counting script. 

Known Issues:
It works well with normal photos, however. If there is a photo with a lot of background noise, the script will pick up more particles than there actually are. This is a thresholding issue. We are currently using the Moments auto-thresholding method, which is good for most of the pictures. 

TODO:
Find a better thresholding strategy. 
Add a dialogue box for other users to choose the parameters they want to use.